% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{create_pull_request}
\alias{create_pull_request}
\title{Creates a pull request in the specified repository}
\usage{
create_pull_request(title, targets, description = NULL,
  clientRequestToken = NULL)
}
\arguments{
\item{title}{The title of the pull request. This title will be used to identify the pull request to other users in the repository.}

\item{targets}{The targets for the pull request, including the source of the code to be reviewed (the source branch), and the destination where the creator of the pull request intends the code to be merged after the pull request is closed (the destination branch).}

\item{description}{A description of the pull request.}

\item{clientRequestToken}{A unique, client-generated idempotency token that when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request will return information about the initial request that used that token.

The AWS SDKs prepopulate client request tokens. If using an AWS SDK, you do not have to generate an idempotency token, as this will be done for you.}
}
\description{
Creates a pull request in the specified repository.
}
\examples{

}
