% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{create_repository}
\alias{create_repository}
\title{Creates a new, empty repository}
\usage{
create_repository(repositoryName, repositoryDescription = NULL)
}
\arguments{
\item{repositoryName}{The name of the new repository to be created.

The repository name must be unique across the calling AWS account. In addition, repository names are limited to 100 alphanumeric, dash, and underscore characters, and cannot include certain characters. For a full description of the limits on repository names, see \href{http://docs.aws.amazon.com/codecommit/latest/userguide/limits.html}{Limits} in the AWS CodeCommit User Guide. The suffix ".git" is prohibited.}

\item{repositoryDescription}{A comment or description about the new repository.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.}
}
\description{
Creates a new, empty repository.
}
\examples{

}
