% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{delete_branch}
\alias{delete_branch}
\title{Deletes a branch from a repository, unless that branch is the default branch for the repository}
\usage{
delete_branch(repositoryName, branchName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the branch to be deleted.}

\item{branchName}{[required] The name of the branch to delete.}
}
\description{
Deletes a branch from a repository, unless that branch is the default branch for the repository.
}
\section{Accepted Parameters}{
\preformatted{delete_branch(
  repositoryName = "string",
  branchName = "string"
)
}
}

