% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{delete_comment_content}
\alias{delete_comment_content}
\title{Deletes the content of a comment made on a change, file, or commit in a repository}
\usage{
delete_comment_content(commentId)
}
\arguments{
\item{commentId}{[required] The unique, system-generated ID of the comment. To get this ID, use GetCommentsForComparedCommit or GetCommentsForPullRequest.}
}
\description{
Deletes the content of a comment made on a change, file, or commit in a repository.
}
\section{Accepted Parameters}{
\preformatted{delete_comment_content(
  commentId = "string"
)
}
}

