% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{delete_repository}
\alias{delete_repository}
\title{Deletes a repository}
\usage{
delete_repository(
  repositoryName = "string"
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to delete.}
}
\description{
Deletes a repository. If a specified repository was already deleted, a null repository ID will be returned.
}
\details{
Deleting a repository also deletes all associated objects and metadata. After a repository is deleted, all future push calls to the deleted repository will fail.
}
\examples{

}
