% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_blob}
\alias{get_blob}
\title{Returns the base-64 encoded content of an individual blob within a repository}
\usage{
get_blob(repositoryName, blobId)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the blob.}

\item{blobId}{[required] The ID of the blob, which is its SHA-1 pointer.}
}
\description{
Returns the base-64 encoded content of an individual blob within a repository.
}
\section{Accepted Parameters}{
\preformatted{get_blob(
  repositoryName = "string",
  blobId = "string"
)
}
}

