% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_branch}
\alias{get_branch}
\title{Returns information about a repository branch, including its name and the last commit ID}
\usage{
get_branch(repositoryName = NULL, branchName = NULL)
}
\arguments{
\item{repositoryName}{The name of the repository that contains the branch for which you want to retrieve information.}

\item{branchName}{The name of the branch for which you want to retrieve information.}
}
\description{
Returns information about a repository branch, including its name and the last commit ID.
}
\section{Accepted Parameters}{
\preformatted{get_branch(
  repositoryName = "string",
  branchName = "string"
)
}
}

