% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_comment}
\alias{get_comment}
\title{Returns the content of a comment made on a change, file, or commit in a repository}
\usage{
get_comment(
  commentId = "string"
)
}
\arguments{
\item{commentId}{[required] The unique, system-generated ID of the comment. To get this ID, use GetCommentsForComparedCommit or GetCommentsForPullRequest.}
}
\description{
Returns the content of a comment made on a change, file, or commit in a repository.
}
\examples{

}
