% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_commit}
\alias{get_commit}
\title{Returns information about a commit, including commit message and committer information}
\usage{
get_commit(repositoryName, commitId)
}
\arguments{
\item{repositoryName}{The name of the repository to which the commit was made.}

\item{commitId}{The commit ID. Commit IDs are the full SHA of the commit.}
}
\description{
Returns information about a commit, including commit message and committer information.
}
\examples{

}
