% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_file}
\alias{get_file}
\title{Returns the base-64 encoded contents of a specified file and its metadata}
\usage{
get_file(
  repositoryName = "string",
  commitSpecifier = "string",
  filePath = "string"
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the file.}

\item{commitSpecifier}{The fully-quaified reference that identifies the commit that contains the file. For example, you could specify a full commit ID, a tag, a branch name, or a reference such as refs/heads/master. If none is provided, then the head commit will be used.}

\item{filePath}{[required] The fully-qualified path to the file, including the full name and extension of the file. For example, /examples/file.md is the fully-qualified path to a file named file.md in a folder named examples.}
}
\description{
Returns the base-64 encoded contents of a specified file and its metadata.
}
\examples{

}
