% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_folder}
\alias{get_folder}
\title{Returns the contents of a specified folder in a repository}
\usage{
get_folder(
  repositoryName = "string",
  commitSpecifier = "string",
  folderPath = "string"
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository.}

\item{commitSpecifier}{A fully-qualified reference used to identify a commit that contains the version of the folder's content to return. A fully-qualified reference can be a commit ID, branch name, tag, or reference such as HEAD. If no specifier is provided, the folder content will be returned as it exists in the HEAD commit.}

\item{folderPath}{[required] The fully-qualified path to the folder whose contents will be returned, including the folder name. For example, /examples is a fully-qualified path to a folder named examples that was created off of the root directory (/) of a repository.}
}
\description{
Returns the contents of a specified folder in a repository.
}
\examples{

}
