% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_pull_request}
\alias{get_pull_request}
\title{Gets information about a pull request in a specified repository}
\usage{
get_pull_request(
  pullRequestId = "string"
)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID of the pull request. To get this ID, use ListPullRequests.}
}
\description{
Gets information about a pull request in a specified repository.
}
\examples{

}
