% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_repository}
\alias{get_repository}
\title{Returns information about a repository}
\usage{
get_repository(repositoryName)
}
\arguments{
\item{repositoryName}{The name of the repository to get information about.}
}
\description{
Returns information about a repository.
}
\details{
The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.
}
\examples{

}
