% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{get_repository_triggers}
\alias{get_repository_triggers}
\title{Gets information about triggers configured for a repository}
\usage{
get_repository_triggers(repositoryName)
}
\arguments{
\item{repositoryName}{The name of the repository for which the trigger is configured.}
}
\description{
Gets information about triggers configured for a repository.
}
\examples{

}
