% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{list_branches}
\alias{list_branches}
\title{Gets information about one or more branches in a repository}
\usage{
list_branches(
  repositoryName = "string",
  nextToken = "string"
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the branches.}

\item{nextToken}{An enumeration token that allows the operation to batch the results.}
}
\description{
Gets information about one or more branches in a repository.
}
\examples{

}
