% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{list_repositories}
\alias{list_repositories}
\title{Gets information about one or more repositories}
\usage{
list_repositories(nextToken = NULL, sortBy = NULL, order = NULL)
}
\arguments{
\item{nextToken}{An enumeration token that allows the operation to batch the results of the operation. Batch sizes are 1,000 for list repository operations. When the client sends the token back to AWS CodeCommit, another page of 1,000 records is retrieved.}

\item{sortBy}{The criteria used to sort the results of a list repositories operation.}

\item{order}{The order in which to sort the results of a list repositories operation.}
}
\description{
Gets information about one or more repositories.
}
\examples{

}
