% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_package.R
\docType{package}
\name{paws.codecommit-package}
\alias{paws.codecommit}
\alias{paws.codecommit-package}
\title{paws.codecommit: AWS CodeCommit}
\description{
This is the *AWS CodeCommit API Reference*. This
reference provides descriptions of the operations and data types for
AWS CodeCommit API along with usage examples.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
