% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{post_comment_for_compared_commit}
\alias{post_comment_for_compared_commit}
\title{Posts a comment on the comparison between two commits}
\usage{
post_comment_for_compared_commit(
  repositoryName = "string",
  beforeCommitId = "string",
  afterCommitId = "string",
  location = list(
    filePath = "string",
    filePosition = 123,
    relativeFileVersion = "BEFORE"|"AFTER"
  ),
  content = "string",
  clientRequestToken = "string"
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository where you want to post a comment on the comparison between commits.}

\item{beforeCommitId}{To establish the directionality of the comparison, the full commit ID of the 'before' commit.}

\item{afterCommitId}{[required] To establish the directionality of the comparison, the full commit ID of the 'after' commit.}

\item{location}{The location of the comparison where you want to comment.}

\item{content}{[required] The content of the comment you want to make.}

\item{clientRequestToken}{A unique, client-generated idempotency token that when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request will return information about the initial request that used that token.}
}
\description{
Posts a comment on the comparison between two commits.
}
\examples{

}
