% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{post_comment_reply}
\alias{post_comment_reply}
\title{Posts a comment in reply to an existing comment on a comparison between commits or a pull request}
\usage{
post_comment_reply(inReplyTo, content, clientRequestToken = NULL)
}
\arguments{
\item{inReplyTo}{The system-generated ID of the comment to which you want to reply. To get this ID, use GetCommentsForComparedCommit or GetCommentsForPullRequest.}

\item{content}{The contents of your reply to a comment.}

\item{clientRequestToken}{A unique, client-generated idempotency token that when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request will return information about the initial request that used that token.}
}
\description{
Posts a comment in reply to an existing comment on a comparison between commits or a pull request.
}
\examples{

}
