% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{put_repository_triggers}
\alias{put_repository_triggers}
\title{Replaces all triggers for a repository}
\usage{
put_repository_triggers(
  repositoryName = "string",
  triggers = list(
    list(
      name = "string",
      destinationArn = "string",
      customData = "string",
      branches = list(
        "string"
      ),
      events = list(
        "all"|"updateReference"|"createReference"|"deleteReference"
      )
    )
  )
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository where you want to create or update the trigger.}

\item{triggers}{[required] The JSON block of configuration information for each trigger.}
}
\description{
Replaces all triggers for a repository. This can be used to create or delete triggers.
}
\examples{

}
