% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{test_repository_triggers}
\alias{test_repository_triggers}
\title{Tests the functionality of repository triggers by sending information to the trigger target}
\usage{
test_repository_triggers(repositoryName, triggers)
}
\arguments{
\item{repositoryName}{The name of the repository in which to test the triggers.}

\item{triggers}{The list of triggers to test.}
}
\description{
Tests the functionality of repository triggers by sending information to the trigger target. If real data is available in the repository, the test will send data from the last commit. If no data is available, sample data will be generated.
}
\examples{

}
