% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{update_default_branch}
\alias{update_default_branch}
\title{Sets or changes the default branch name for the specified repository}
\usage{
update_default_branch(repositoryName, defaultBranchName)
}
\arguments{
\item{repositoryName}{The name of the repository to set or change the default branch for.}

\item{defaultBranchName}{The name of the branch to set as the default.}
}
\description{
Sets or changes the default branch name for the specified repository.
}
\details{
If you use this operation to change the default branch name to the current default branch name, a success message is returned even though the default branch did not change.
}
\examples{

}
