% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codecommit_operations.R
\name{update_pull_request_description}
\alias{update_pull_request_description}
\title{Replaces the contents of the description of a pull request}
\usage{
update_pull_request_description(pullRequestId, description)
}
\arguments{
\item{pullRequestId}{The system-generated ID of the pull request. To get this ID, use ListPullRequests.}

\item{description}{The updated content of the description for the pull request. This content will replace the existing description.}
}
\description{
Replaces the contents of the description of a pull request.
}
\examples{

}
